pyproject.toml

[project]
name = "vision-app"
version = "1.0.0"
description = "Tkinter + OpenCV app"
requires-python = ">=3.9"
dependencies = [
  "opencv-python>=4.9",
  "pillow>=10.0",
  "numpy>=1.26",
  "shapely>=2.0",
]

[tool.ruff]
line-length = 100
select = ["E","F","I","UP"]
